// apps/marketplace/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
 
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ERole = {
  OWNER: 'OWNER',
  COURIER: 'COURIER',
  CUSTOMER: 'CUSTOMER'
} as const

export type ERole = (typeof ERole)[keyof typeof ERole]


export const EVoucherCategory = {
  FOOD: 'FOOD',
  MEDICINE: 'MEDICINE',
  TRANSPORT: 'TRANSPORT'
} as const

export type EVoucherCategory = (typeof EVoucherCategory)[keyof typeof EVoucherCategory]


export const EProductType = {
  EDIBLE: 'EDIBLE',
  NON_EDIBLE: 'NON_EDIBLE'
} as const

export type EProductType = (typeof EProductType)[keyof typeof EProductType]


export const EVehicleType = {
  BIKE: 'BIKE',
  SCOOTER: 'SCOOTER',
  CAR: 'CAR',
  ON_FOOT: 'ON_FOOT'
} as const

export type EVehicleType = (typeof EVehicleType)[keyof typeof EVehicleType]


export const EOrderStatus = {
  PENDING_PAYMENT: 'PENDING_PAYMENT',
  PENDING_OWNER: 'PENDING_OWNER',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
  READY: 'READY',
  TAKEN: 'TAKEN',
  PICKED_UP: 'PICKED_UP',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED'
} as const

export type EOrderStatus = (typeof EOrderStatus)[keyof typeof EOrderStatus]